ECS added into project with testing.
Added ECS with testing:
Entity Manager to create entities
Component Manager to create components
Component Array to hold the array of components - this is using a vector at the moment and could be explored later to improve. Idea is that we could look at making a build system that reserves/predicts how many components, this way we don't waste memory over reserving.