Scene Serialization with member variables
Kept scene serialization method (to string) in own utils class as adding in classes such as Vec3 add extra complexity to runtime when its not needed unless its ifdefed out. Additionally it adds a level of abstraction on top of glm::vec3 which might optimise out some of the methods for.
More thought needs to be put into vec3 and vec4 before I put it in those classes