Added textures, more robust obj loading and Filepath class
Filepath class is a helper class that is added to allow filepaths to be loaded in the project from the project directory.

Robust obj loading to allow for loading of obj with normals, texture coordinates and indices. Added error checking if any aren't present, as well as warnings for new keywords from the obj files.

Added textures for instances. This just makes use of the descriptor sets from the imgui pool, so this will have to be architected better in the future, but is a good starting point to keep the textures in. Textures can also be loaded from disk

Removed vulkan bootstrap as it wasn't used.