Material Manager added, start on creating assets, ULID class added
Added Material Manager for the engine and Editor
Current pathing is that the material manager will create materials, whilst the editor version will do this whilst keeping some more information such as all the material info of textures loaded
ULID class added for unique ID's of instances
Start on creating assets via right-click context menu. At the moment, this just creates a folder with a .scarlett_asset extension however needs to be expanded to create a default asset and serialize it.